requirements.txt
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
arxiv==2.1.0
attrs==25.3.0
Authlib==1.5.2
babel==2.17.0
beautifulsoup4==4.13.4
bibtexparser==1.4.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
dnspython==2.7.0
docutils==0.21.2
dotenv==0.9.9
duckduckgo_search==7.5.1
exceptiongroup==1.2.2
fake-useragent==2.2.0
feedparser==6.0.10
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
free-proxy==1.1.3
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-genai==1.0.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lxml==5.3.0
MarkupSafe==3.0.2
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.1
plotly==5.22.0
primp==0.14.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.8.0
pymongo==4.6.2
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scholarly==1.7.11
scikit-learn==1.4.0
scipy==1.13.1
selenium==4.31.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.7
Sphinx==7.4.7
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.2
Werkzeug==2.3.7
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.2.0
